af-main: Refactor of user session management
This changes how user session are started and handled
within systemd. This evolution add a tiny program to
start and handle user sessions: afm-user-session.

This is the first step of a serie because, actually,
due to platform restrictions, the common users if started
this way will not be able to run correctly (issues with
wayland/weston and with handling of CGROUP2 with Smack LSM)

Also fixes memory leaks and improves build process.

Bug-AGL: SPEC-545
Bug-AGL: SPEC-1016

Change-Id: Icaf7c575633fc663a3df0ffee683b22cc087bb16
Signed-off-by: José Bollo <jose.bollo@iot.bzh>