gstreamer: add libv4l2 support
libv4l2 support is useful for getting more out of v4l2src. I discovered
this when trying to get RGB format video out of a camera and seeing that
libv4l2 can emulate it (even when the camera does not support it).
However, that emulation must be compiled into gstreamer, so let's do
that.

Change-Id: I1117277dd48fefbbe2a9a1e80f274cb145f3a5e5
Signed-off-by: Martin Kelly <mkelly@xevo.com>