dbus-cynara: Fix a missing RDEPENDS
dbus-cynara is a separate package of dbus because it
allows to break the dependency loop dbus -> cynara -> ... -> dbus
coming from the fact that many many usefull things depend
on dbus: documentation generators, test handlers, ...

In other words, dbus-cynara is the same as dbus. As such, it
uses the subpackage dbus-lib (known as libdbus). This has to
be set as a RDEPENDS, otherwise bitbake complains:

 QA Issue: dbus-cynara rdepends on dbus-lib, but it isn't a builds
           dependency, missing dbus in DEPENDS or PACKAGECONFIG?

Change-Id: I72472dc9e6e8f21d0aabc9a1186f1cb7d8343445
Signed-off-by: José Bollo <jose.bollo@iot.bzh>