Introduce meta-agl-profile-core and meta-agl-profile-graphics
Rework towards agl profiles.

This change is part of a series of changes to create the
AGL profiles.

This set will mainly introduce the 'core' profile.

It is setup to be a drop-in change, thus some files were kept
in (dummy) locations for now.
However, they'll be taken care of in the next changes in this series.

The main target of the meta-agl-profile-core layer is to host:
- a minimal, bootable image with network and package management enabled
  -- agl-image-boot
- a minimal image with network and packagemanagement and the AGL APIs
  -- agl-image-minimal

The layer meta-agl-profile-graphical is used as superset of these
and includes support for egl+wayland+weston. All recipes concerning
graphics were moved there. This is not a full profile as we still have
to migrate some parts of meta-agl-demo in a follow-up changeset.

The roadmap as discussed during the F2F session in Karlsruhe is:
- week 16 : core profile and profiles w/o graphics
- week 17 : graphical profiles
- week 18 : final conversion of the demo image

v2: moved agl-login-manager from -graphics to -core (see Jose's comment)
v3: moved back after discussion - follow-up in separate changeset

Change-Id: Idacb0d1274baac1f63f8d1b850d4b1104ac33918
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>