2nd part of the layer/profile rework [1/2]
Main goal of this is to move the agl-service-* bindings into the core and to
unify common recipes like wayland/weston into meta-agl-profile-graphical.
As a placeholder for recipes is meta-agl-profile-demo which are being reshuffled.

TLDR: agl-service move to meta-agl-profile-core. wayland/weston moves to meta-agl-profile-graphical.

This patchset needs the 2nd patchset of the series to build (meta-agl-demo).

Bug-AGL: SPEC-145 , SPEC-876 , SPEC-877

v2: rebase including udiscs patches from meta-agl-demo

Change-Id: I5e33f72d01707fa8b826d382bccaca3addcbbc47
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>