Add machine templates for qemuarm and qemuarm64
This changeset adds the necessary tepmlates to build the qemu
targets for ARM 32bit and 64bit.

The MACHINE names are 'qemuarm' and 'qemuarm64'.

As the compiler flags for qemuarm in Yocto are using the equivalent
of armv5te, we need to ship our own machine/qemuarm.conf with our
default tuning for armv7-a (YP name armv7ve), and overrides to build
a v7 kernel and use a v7 capable QEMU machine.  For qemuarm64, some
additional kernel configuration is enabled to allow building and
loading of the MOST drivers.

Bug-AGL: SPEC-756

Change-Id: I0a688a1d61a75f1287902fc2733a8a54b0485a34
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Signed-off-by: Scott Murray <scott.murray@konsulko.com>