Fix cmake build with sdk inside a script
For the sdk build the cmake config file is source:
sysroots/x86_64-aglsdk-linux/environment-setup.d/cmake.sh

But use "alias" :
alias cmake="cmake ..."

Unfortunatly alias is not expand inside a script.

using an environnement variable for "CMAKE_TOOLCHAIN_FILE" solve the
issue.

This Fix is a port of the yocto FIX 139551:
https://patchwork.openembedded.org/patch/139551/
Unfortunatly this patch is not include in yocto upstream

Bug-AGL: SPEC-1530

Change-Id: Ica80f66cbe976a6ccb1acb7d96b4c4ce4252d4f3
Signed-off-by: Ronan Le Martret <ronan.lemartret@iot.bzh>