af-main: set relax handling for apps signatures even when agl-devel is not set
Allow widgets to be unsigned even in non-development mode. This is a temporary
workaround to let some time to setup the apps signing workflow.

This workaround can be disabled for smoother transition to the world where
all apps must be signed. For this, define AGL_FORBID_UNSIGNED_APPS="1" in
local.conf.

**IMPORTANT: THIS PATCH MUST BE REVERTED IN THE FUTURE AND ALL APPS MUST BE SIGNED**

This patch alters the correct behaviour that was introduced here:
https://gerrit.automotivelinux.org/gerrit/#/c/15527/1/meta-app-framework/recipes-core/af-main/af-main_1.0.bb

Bug-AGL: SPEC-1614

Change-Id: I6bf50b70905a00445598f42ef5f6d9326c6d4cfe
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>