linux-firmware: Enable WiFi for Raspberry Pi 3 B+
Add binary blobs for brcmfmac43455 to enable WiFi for the new
Raspberry Pi 3 Model 3 B+. At the moment meta-raspberrypi
provides this for latest releases but not for Rocko. This fix
should remain in meta-agl until AGL moves to sumo or until we
backport the required changes to meta-raspberrypi branch rocko.

Bug-AGL: SPEC-1652

Change-Id: I644938ac3799b0e831f022fff9b2bc85aa8fd27c
Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>