Make it possible to use systemd-coredump for development
For development, systemd-coredump should be enabled especially
for meta-agl/meta-agl-demo because all applications which
compatible with agl appfw are running on systemd.

This patch make systemd-coredump available when
agl-devel is set on (run aglsetup.sh with 'agl-devel').

The core dump files are stored in /var/lib/systemd/coredump/
and 'coredumpctl' is helpful to use dump files.

Change-Id: Ic8a011e96fe36be0266389c03d761d1b8e78bf68
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>