Change the default sdk install path
Installing 2 SDKs of the same version but different arch
results in broken compilation due to a bug in the meta-qt5 sdk
environment scripts.

Mitigate it by adding the DEFAULTTUNE to the SDK installation
folder until the issue can be addressed in SPEC-1686.

Bug-AGL: SPEC-1667, SPEC-1686

v2 (jsmoeller): rewrite to use /opt and SDKPATH

Change-Id: Id7e6c919631fc0bcc630a6b813a714f83adaf148
Signed-off-by: Ronan Le Martret <ronan.lemartret@iot.bzh>
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>