meta-agl-bsp: meta-ti: amx3-cm3: Add a fix for security_flags.inc
The way that the PM firmware for these devices is built takes very
specific steps to only get the cross-compiler provided by OE and almost
nothing else.  Unfortunately when we build with the security_flags.inc
file in our distro this changes some of the default behavior of the
compiler and in turn we need to make sure to disable this and go back to
the default behavior.

Change-Id: I1ab8e20f77d7ab6a3c786437e9a9422b24f0e159
Bug-AGL: SPEC-1341
Signed-off-by: Tom Rini <trini@konsulko.com>