meta-agl-bsp: kernel: fix can-bus.cfg
This config fragment for CAN bus had both an option for enabling as a
module:

  CONFIG_CAN_C_CAN=m

and for disabling:

  # CONFIG_CAN_C_CAN is not set

Since the disable came last, disable wins.  This caused the C_CAN
driver and the C_CAN_PLATFORM driver (used by the TI dra7-evm) to be
disabled in the kernel.

To fix, remove the option to disable.

Change-Id: I5be0cee161318cdfe7deb9c5e9c9f15d79cda7ac
Fixes: d870883e7b29d212dc3e2a0e6158d837d4356c6a
Signed-off-by: Kevin Hilman <khilman@baylibre.com>