libmicrohttpd: Fix listening and epoll together
It appeared recently that libmicrohhtpd isn't able
to manage listening interfaces and at the same time
to manage in/out streams using epoll.

After discussion with maintainers of libmicrohttpd
this is a bug and is already corrected. See:

 https://gnunet.org/git/libmicrohttpd.git/commit/?id=3fc686e

While waiting for integration of this patched version
in poky/openembedded, it is needed to have that patch
integrated in AGL for running binders aware of listening
interfaces.

Upstream-Status: Accepted

Bug-AGL: SPEC-1896

Change-Id: Ib436f6b4be27e638fedbae22f690109fb064c8e7
Signed-off-by: José Bollo <jose.bollo@iot.bzh>