create-combined-dtb: Combine dtb and dtbo
Combine a number of device tree overlays to a base blob with a
specified filename in the deploydir only. This feature is
required only for netboot therefore the files are not included
in the image.

Bug-AGL: SPEC-1775

Change-Id: I16ff043ee0d44d4151e1fb2ba714bd1542854bd8
Suggested-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>
(cherry picked from commit 9cffd21d1ec735c871943a301d46e3816e18a320)