Add instrument cluster profile definitions
Add definitions for instrument cluster profiles for plain Wayland/Weston
based images and images that also include Qt5 support.  Due to the
current requirement that profiles only be a superset of one parent
profile, the Qt5 profile inherits from agl-profile-graphical-qt5 and not
agl-profile-cluster as one might expect.  This is due to some of the
required recipes and bbappends for Qt5 usage only being available in the
meta-agl-profile-graphical-qt5 layer.

Change-Id: I863b35bdc009775ac204a720bbc2cc85e71c268a
Signed-off-by: Scott Murray <scott.murray@konsulko.com>