weston-ini-conf: update gst-record logic
The logic to handle the "virtual-output" and "gst-record"
DISTRO_FEATURES seems to be unfinished, this update replaces the
"virtual-output" usage with just "gst-record", as the former does not
seem to exist in the AGL codebase.  The virtualoutput.cfg file has been
updated with the arguments from the virtual output section in the
replacement .ini file from the mapviewer-demo recipe in meta-agl-demo,
but some further investigation is required into tweaking things to
produce a final .ini file that matches that one.  These changes are a
first step to getting to that end goal.

Change-Id: Ib99e0717ed1e66016c6e936a3c398addcb6adfb7
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
(cherry picked from commit 262b016588cb153520362cf0aed3715e43d86672)