weston: add weston-ready
Weston seems to take variable amounts of time to initialize on startup
depending on configuration (e.g. gst-record feature being enabled for
cluster support).  To avoid windowmanager and its dependencies failing,
patch LayerManagerControl so that errors can be detected; then use it
in a new script, weston-ready, to detect if Weston is functional.  A
corresponding systemd service unit is also added so that things can
use it in their dependencies.

Change-Id: I5711e7f9d9a891e1c7a160b5d65711bd6e7fee03
Signed-off-by: Scott Murray <scott.murray@konsulko.com>