connman: Tweak main.conf for cluster demo support
When running the cluster demo, we do not want the second ethernet
interface managed by connman, as it will be used for a separate
connection to the cluster board, and should not be considered part
of the shared connection pool.  To do this, NetworkInterfaceBlacklist
is tweaked if agl-cluster-demo-support is enabled.

Change-Id: Iab8c17662bb7f5b06b5eda6c97dd32a6c5ef3f5a
Signed-off-by: Scott Murray <scott.murray@konsulko.com>