bump bluez-alsa to master (brings Ofono support)
This takes the mandatory commits to have HFP support over Ofono
(56fcc3 from 2018-12-04)
Also adds the needed options to the daemon in the service file
The license changed due to change of year and author list

It was bogus to have hcitop in PACKAGECONFIG.
Namely, this is not a feature in yocto, but only an optional
feature of bluez-alsa (that helps to perform bandwith measures)
However, Ofono is one.

This also adds hcitop and rfcomm tools in development images.
Those tools help to diagnose bt issues.

Change-Id: I426dbee5aa8911c9c64f4ec7399528d149dc651b
Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>