Update weston-ivi-extension
The version of weston-ivi-extension that was added to the
meta-agl-profile-graphical layer in commit 7b56b77 (a cherry-pick from
flounder) missed some updates that were made to the copy in meta-agl-demo
in master after flounder was branched.  Update the patches and recipe to
match the current version from meta-agl-demo so the latter can be removed.

Bug-AGL: SPEC-1730

Change-Id: I8ee2e68679a74b5dbc994ff977c11fa78e9c75b2
Signed-off-by: Scott Murray <scott.murray@konsulko.com>