systemd: Refactor build using smack-system-setup
This changes introduces the new recipe

  meta-security/recipes-core/smack-system-setup/smack-system-setup_1.bb

The purpose is to split the recipe of systemd in two
parts:

 - A part specific to systemd and only systemd
   It actually includes Smack patches for systemd
   and a renaming of udev-rules.

 - A part more oriented on putting the system in
   order to run with Smack activated.

At the end, it will probably save many rebuilds as
systemd recipe will evolve less in relation with the
setup of the system.

As example, the udev rule file "55-udev-smack-default.rules"
that setup udev rules specific to smack is no more brought
by systemd but by smack-system-setup.

Also at the same time, some cleanup and refactoring is
done. Note that the ".bbappend" file for systemd is
now fixed in version and is including a common file file
that records the several known versions. No cleanup was
made on the versioned patch for the sake of memory.
The cleanup of the history is to be achieved later...

Bug-AGL: SPEC-2045

Change-Id: Iacf772142a381729dfdbe98d133a3effc4d6cf68
Signed-off-by: José Bollo <jose.bollo@iot.bzh>