connman+bluez5: Update rights for smack systems
Reading the file /etc/resolv.conf that is linked to
/run/connman/resolv.conf is not possible for common users.

This changes add the setting of the directory /run/connman
that allows common applications to read that file.

To achieves this goal, that changes use the intended
tuning mechanism of systemd instead of using sed.
This is cleaner. Thus this as been adapted for bluez5 too.

Bug-AGL: SPEC-2006

Change-Id: I3d2a708be2a5c62664bfcf90757e9e5c080d6179
Signed-off-by: José Bollo <jose.bollo@iot.bzh>