connman: Ignore MOST ethernet interfaces
On systems like the CES demo units with MOST adapters, the virtual
ethernet interface created by MOST interferes with connman's
management of the shared connection of the Wi-fi and main ethernet
interface, eg. the MOST ethernet being present results in the Wi-fi
not auto-connecting at boot.  To avoid this, ignore the "meth"
interfaces by adding them to the blacklist.  This seems reasonable,
as they are unlikely to be used for internet connectivity.

Change-Id: I9b56333df4b4d7e128a765855065efce58abb734
Signed-off-by: Scott Murray <scott.murray@konsulko.com>