AGL Layer Overview: Re-write of the README-AGL.md file
This topic is really about overviewing the top-level Yocto
layers that are part of the AGL build system.  These layers
are meta-agl, meta-agl-demo, and meta-agl-devel.  The
README-AGL.md file now simply introduces this fact and
is used as a reading guide for the subsequent layer
topics.  All AGL UCB type information previously in the
file is redundant for the docs.

Change-Id: Iaeb8d192db977ef47f2f1075de5025acf6fe7019
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>