Signalize run-agl-postinsts error appropriately to systemd
If any of the post-install scripts fail during first boot,
they should be retried on subsequent boots.

That was not happening because run-agl-postinsts.service
would be disabled even if a post-install script failed as
it always returned success on its execution.

Now signalize properly when it sees any error while executing
some post-install script.

Change-Id: I4402ee69c80814b06e3cc25cb400a66de3b5652d
Signed-off-by: Andre Goddard Rosa <andre.goddard@gmail.com>