neardal: lib: fix random memory corruption
This commit includes a neardal patch to fix
the random segfault seen on agl-service-nfc.

The problem (random memory corruption)
was due to invalid access to freed
memory on neardal library.

The proposed fix has been submitted upstream
and this commit is only a temporary measure
while the solution is not provided directly from
neardal repo.

Bug-AGL: SPEC-1976
Change-Id: I21984cb8135537ff1232a4387a31688e1a140642
Signed-off-by: raquel medina <raquel.medina@konsulko.com>