neard: cancel empty record processing on tag detection
neard service clients registered for tag interface call a series
of getter methods via dbus. In the case of a tag programmed with a
valid NDEF containing an empty record (D0 00 00) no valid dbus
string type is associated with the record and the
get_type_property call results in a dbus error and neard service
exit.

This commit offers a workaround to avoid this particular situation
while a more comprehensive solution is found (there are other
record types that would fall in this same scenario).

Bug-AGL: SPEC-2123
Change-Id: I1b6d639bdc2aea0a331e8c06d6eac18da21482f5
Signed-off-by: raquel medina <raquel.medina@konsulko.com>