meta-agl-bsp: recipes-kernel: linux: Increase BLK_RAM size
For SPEC-2075, I tried to boot an upsquare using NBD and it fail with:
[    8.380343] EXT4-fs (ram0): bad geometry: block count 8492 exceeds size of device (4096 blocks)
[    8.390509] EXT4-fs (ram0): bad geometry: block count 8492 exceeds size of device (4096 blocks)

For having a successfull boot, I needed to increase BLK_DEV_RAM_SIZE to
a larger value (16384).
Instead of increasing to only 16384, I use 65536 which is the value used in
other meta-xxxx like meta-freescale.

This patch increase BLK_DEV_RAM_SIZE to 65536.

Signed-off-by: Corentin LABBE <clabbe@baylibre.com>
Change-Id: I0792be1038d0bcd120c3c69c90cd6e1966c9b1ee