Fix imx6qdlsabreauto machine configuration
The imx6qsabreauto target was falling behind and could not be build for GG-7.0.

The commit 808fff5 from 'AGL-repo' removed reference to external repos meta-freescale-3rdparty and meta-freescale-distro.
However 'meta-agl' was not updated to follow up on this change which broke configuration for imx6 based targets.

This patch makes the required changes to fix the imx6qsabreauto machine configuration:

    - removes reference to deleted layers: meta-freescale-3rdparty and meta-freescale-distro
    - updates MACHINE value to imx6qdlsabreauto as defined in meta-freescale layer
    - deletes reference to packagegroup-fscl-* not required to build AGL
    - add "agl-medium-arm-compiler" to DISTRO_FEATURES so DEFAULTTUNE is correctly set to 'armv7athf-neon'

v2 (jsmoeller): moved board name to match MACHINE

v3 (bechir.mghirbi):
    - remove wayland_2.0.0 as weston_4.0.0.imx is used.
    - use default IMAGE_FSTYPES as bitbake is complaining that no IMAGE_CMD defined for IMAGE_FSTYPES entry 'sdcard'
    - Add CFG80211 Kernel configuration

Change-Id: Ib8b1d254972c737725247a5b13ce333e8ef9d4dc
Signed-off-by: Bechir Mghirbi <bechir.mghirbi@grammer.com>
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>