af-main: Add ability to manage all widget
For system uses -like automatic testing-, it is
needed to include tha ability to list ALL installed
widgets and ALL running instance of installed
widgets.

This change adds the option -a or --all to the
commands 'list' and 'runners' for the purpose
of listing any installed widgets, even those
requiring to be hidden.

Example: afm-util ps -a

Bug-AGL: SPEC-2272
Bug-AGL: SPEC-1850

Change-Id: I7a6896e79fa624dee7ee1f18d1e75c14f61bed0c
Signed-off-by: José Bollo <jose.bollo@iot.bzh>