meta-agl-bsp: meta-sancloud: weston-ini-conf: Set correct gbm-format
The default gbm-format needs to be changed on the BeagleBone Enhanced
due to a hardware errata on the AM335x SoC.

For further details see the following section of the TI wiki:
http://processors.wiki.ti.com/index.php/Processor_Linux_SDK_Graphics_and_Display#AM3_Beagle_Bone_Black_Board_Configuration

The TI documentation recommends further changes to powervr.ini and
qt_env.sh. The change to powervr.ini is already applied in the
meta-sancloud layer and the change to qt_env.sh was found not to be
needed.

Bug-AGL: SPEC-2409

Change-Id: Iec91ab939b13bd4f936e64681d45931dafb8ee91
Signed-off-by: Paul Barker <paul.barker@sancloud.co.uk>