meta-agl-profile-core: Add mosquitto to passwd / group
Add entries for the mosquitto daemon to the static passwd and group
files to allow building it into images.  Since there is not really a
mechanism for doing this in the telematics profile without a full
duplication of the files, adding it to the core version seems required
for now.

Change-Id: I50248a7d2b2a2ed5539ba002d9521c0e0fd14034
Signed-off-by: Scott Murray <scott.murray@konsulko.com>