weston: Remove user session login for weston
Before that commit, the service weston was started as
user 'display' with a full setup of the user.
It means that all services normally started for user
started for weston: systemd-user, dbus, afm-user-daemon,
lightmediascanner (that is all at the moment). This is
useless and grabs resources.

This commit avoids to start a session for weston.
This saves resources. This changes is part of a set
of changes to run services as not root.

Bug-AGL: SPEC-545
Bug-AGL: SPEC-2492

Change-Id: If51d8fdf8c5e19d10151a2b96a79dc01bbb8ccb4
Signed-off-by: José Bollo <jose.bollo@iot.bzh>