af-main: Add dependency to m4 and sed
The file /etc/afm-unit.conf was genrated from
temporary files manually generated. This was
inefficient and error prone.

This was reworked to enable a straight forward
process that directly calls m4 and sed in the
CMAKE files.

Current commit adapt the recipe to that new process
by adding the new build dependencies.

Bug-AGL: SPEC-2436

Change-Id: I662f7620beccc53b6abe58c14c82ac94ae818491
Signed-off-by: José Bollo <jose.bollo@iot.bzh>