setup_mm_packages.sh: Use GFX 3.19 binaries instead of 3.15
Renesas BSP 3.19 is not actually available. But the graphics binaries are.
So this patch get the 3.19 GFX binaries to patch the 3.15 GFX binaries ones.
To achieve that, the following zip is needed:
"R-Car_Gen3_Series_Evaluation_Software_Package_for_Linux-weston5-20190516.zip"
(Note: the name of the zip can change).

By this way, BSP 3.15 of Renesas will be used with the 3.19 GFX binaries.
This solves the graphic issue of the M3ULCB and H3ULCB boards.

This hotfix has been tested on m3ulcb with success.

Warning: This patch has to be revert when the BSP of Renesas will
be updated to 3.19 (at least).

Bug-AGL: SPEC-2253

Change-Id: Ieaab76fb08618f9994c3ff5969d9a361d08b65d2
Signed-off-by: Pierre MARZIN <pierre.marzin@iot.bzh>