lightmediascanner: change db directory
Lightmediascanner runs for one user and should not be able
to look over the the shoulder of other users. So it should
not run as root anymore.

At the moment, lightmediascanner is a user daemon started in the
user session by systemd, without capabilities but with the Smack
label "System".

To allow the service-mediascanner to access its database,
the db must be set in a specific place with the User::App-shared
label.

Bug-AGL: SPEC-1016

Change-Id: Id4546e2567c4f5a0a5139d274d35d11e4879e42e
Signed-off-by: José Bollo <jose.bollo@iot.bzh>