Fix rpi touchscreen support on master
We seem to miss a few bits that changed between kernel 4.9 and kernel 4.14
regarding the 7'' touchscreen.

One addition is the rpi-backlight.dtbo which we also need to use in our create-combined-dtb
recipe for CI use.

We also seem to require https://github.com/raspberrypi/linux/pull/2693#issue-217750943 .

Thanks to Scott Murray for the help digging through this.

v2 and v3: add devicetree changes
v4 and v5: fix name of dtbo
v6: move append to machine include - we need it as a global var
v7: add to config.txt for sdcard boot case

Bug-AGL: SPEC-2465

Change-Id: I2bb3cd974b74a790292e1f36ffdca034928e0bca
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>