Pin SRCREV used by ti-sgx-ddk-km
The commit of the omap5-sgx-ddk-linux git repository used by the
ti-sgx-ddk-km recipe is changed by recent commits to meta-ti to
introduce generic support for armel (soft-float) targets. This conflicts
with the existing patch in meta-agl which adds support for our
toolchain.

The simplest fix is to pin to a working commit by setting SRCREV to the
previous commit on the same branch. This is the same commit that is
currently used and so it won't have any effect until the meta-ti layer
itself is updated.

For reference, the commit in question can be seen here:
https://git.ti.com/graphics/omap5-sgx-ddk-linux/commit/dbdbd9309a58ee723fc827ef5fd45c28347cf61f

Bug-AGL: SPEC-2490

Change-Id: I2924cdadd697680567a9378eb3242f8dd526a429
Signed-off-by: Paul Barker <paul.barker@sancloud.co.uk>