Move the agl-service-radio to the graphical profile
The agl-service-radio relies on gstreamer. Unfortunately the gstreamer recipe is currently
setup to also pull-in wayland (due to a compile-flag if wayland is in DISTRO_FEATURES).

To keep the distro aligned, we don't want to change the DISTRO_FEATURES between the profiles, thus we
move the agl-service radio to the graphical profile where we include weston anyway by default.

For now this means there is no radio service (or mediaplayer service) w/o wayland installed as well.
In case we get this requirement, we'll have to work on the gstreamer recipe and its split.

v2: fix bbapend in meta-agl-bsp/meta-rcar-gen3
v3: put packagegroup-agl-graphical-multimedia.bbappend in graphical profile (not the best place IMHO)
    This fixes the non-graphical build.

Bug-AGL: SPEC-2500

Change-Id: I5ebec7d7713510f3fcd50ae425d7f8cbb6328016
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>