aglsetup: be less restrictive on machines/features definitions
When a private definition for a machine or feature has to be maintained,
the easy way is to have an extra layer 'meta-xxxxxx' containing specific
definitions on top of AGL. This layer can be fetched beside other meta-agl repos.

In the past, aglsetup used to include any definition located in the $METADIR
folder ($METADIR/*/templates/...). With the new layout and change 2b3a44b,
a restriction has been put on folders matching 'meta-ag*', which is very restrictive.

This patch does something in between, by matching all folders 'meta-*'.
NB: with the new layout, this shouldn't include bsp layers or external layers
so it's pretty safe.

Bug-AGL: SPEC-2450

Change-Id: I459ea778380580e5b80685cc86779f36a232ac7c
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>