meta-agl-bsp: Changes to get imx6qdlsabreauto working with etnaviv
Update imx6qdlsabreauto configuration to use open source etnaviv GPU
driver instead of Vivante.  The previously required weston-ini-conf
has been removed as it is not required with etnaviv, and a bbappend
has been added for the linux-fslc kernel that this configuration uses
so it will pull in the AGL kernel bits.  Additionally, the kernel
configuration for linux-fslc is tweaked with config fragments to
better match the other AGL platforms.

Bug-AGL: SPEC-1674, SPEC-2502

Change-Id: Ic40d310b48405ccc7d1435f0fc9f162a129cffe9
Signed-off-by: Scott Murray <scott.murray@konsulko.com>