Add cubox-i and nitrogen6x i.MX6 machines
Add the cubox-i (covering SolidRun Cubox-i, Hummingboard Gate and
Edge), and nitrogen6x (covering Nitrogen 6x and SABRE Lite) machine
definitions.

Note that support for the Broadcom Bluetooth chipset on the SolidRun
i.MX6 SOM is currently missing, as it requires firmware that is not
available in either the linux-firmware package or from the firmware
packages provided by meta-freescale.

Bug-AGL: SPEC-2517

Change-Id: If7bdb9f1af9bb67dba5c1ff836a6d3b04160fefd
Signed-off-by: Scott Murray <scott.murray@konsulko.com>