netboot: systemd must mount /tmp and /run
When using netboot, the initrd script mounts /tmp and /run before pivoting on
the new rootfs and then running systemd. As a consequence, systemd doesn't try
to remount /tmp or /run (and the correct smack labels are not applied on /tmp
and /run, leading to many issues).

This patch unmounts /tmp and /run just before running systemd which will
then mount them with correct labels.

Bug-AGL: SPEC-2596

Change-Id: I4cd7b67f04c232a6e48c777fb5fefa6c46f6744a
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>