meta-rcar-gen3: Add Xen specific device tree for m3ulcb
Provide a specific version of the device tree (appended with -xen) to
run with Xen. Only the R-Car m3ulcb board is supported and tested so
far.

Add also the Xen block device backend support in the kernel, to be
able to run block device based DomU.

From U-Boot, Xen and Linux Dom0 can be run as follows (assuming
binaries are in sdcard first partition):
> setenv bootargs
> fatload mmc 0:1 0x48080000 xen-m3ulcb.uImage
> fatload mmc 0:1 0x48000000 r8a7796-m3ulcb-xen.dtb
> fatload mmc 0:1 0x7a000000 Image-m3ulcb.bin
> bootm 0x48080000 - 0x48000000

The Linux Dom0 bootargs are hardcoded in the device tree, the rootfs
is expected to be on /dev/mmcblk1p2

Bug-AGL: SPEC-2912
Change-Id: Ia86e0789650eaf79d13691cf4f53755c8cb8f06c
Signed-off-by: Jérémy Fanguède <j.fanguede@virtualopensystems.com>