intel-corei7-64: Remove ttyS2 from SERIAL_CONSOLES
Remove ttyS2 from SERIAL_CONSOLES in the intel-corei7-64 include file.
This quiets the log spamming on Up^2 from the device not being present.
The commit that added ttyS2 to SERIAL_CONSOLES in meta-intel (61634f8)
does not explain what boards need it, but since neither the minnowboard
or Up^2 do, not having it by default seems reasonable.

Bug-AGL: SPEC-2678

Change-Id: I9afd1584dd7b92e8d15363c3119b770c592b3c52
Signed-off-by: Scott Murray <scott.murray@konsulko.com>