meta-agl-profile-cluster-qt5: Add qtcompositor-conf recipe
Add qtcompositor-conf recipe to create the udev rule files required to
set up the "display" group ownership of graphics devices when not using
weston.  The recipe is based on what is done in the weston-init bbappend
in the meta-agl-profile-graphical layer.  The package is also added to
the agl-profile-cluster-qtcompositor packagegroup so that the qtcompositor
cluster demo image will pick it up.

Bug-AGL: SPEC-2775

Change-Id: I375e915b2074eb2abcec534a88fddfc6445730e1
Signed-off-by: Scott Murray <scott.murray@konsulko.com>