wayland bug backport: Fix fake "Address already in use" error
Backport a bug modification to the wayland used by AGL,
because the bug may effect the applications start error

The bug modification detail is as bellow:
In the current workflow, socket file will be deleted if it already exists.
However, if the socket file is a symbolic link or the file that it refers
to not exist, we will got "Address already in use", because bind()
thinks the socket file exists and won't create it.
Now, use lstat() to determine whether the socket file exists.

Bug-AGL: SPEC-2783

Signed-off-by: Zhou Mingying <zhoumy@cn.fujitsu.com>
Change-Id: Iff2c7abefa56104eef49ae7102ee8a1de796b813