weston-init: Refactor, simplify, fix
The recipe weston-init was setting smack labels of input devices
to ^. This had the effect to enforce processes that wanted to
read /dev/input/eventXX to have the Smack label System.

I changed the label to * and keep the protection of groups.

At the same time, a refactoring of the recipe make it cleaner to
maintain.

Bug-AGL: SPEC-2796

Change-Id: I3e2345d48a40b15254e68e25bbfcd9b22fcd8629
Signed-off-by: José Bollo <jose.bollo@iot.bzh>