meta-agl-bsp: meta-rcar-gen3-adas: Rework KF radio initialization
Add a systemd service unit and recipe to install it and a udev rules
file so that the Kingfisher Si4689 radio will be initialized on boot
if present, and the I2C device to access it will be accessible to
applications with the "audio" permission.  Note that an accompanying
change to agl-service-radio is required to remove its now unworkable
usage of the si_init script, which is replaced by this change.

Bug-AGL: SPEC-2716

Change-Id: I81cf4b8c439598bdac1eb7910a0d693825141d7b
Signed-off-by: Scott Murray <scott.murray@konsulko.com>