raspberrypi4: Fix WiFi
Apply patch from branch master of meta-raspberrypi as a bbappend
to recipe linux-firmware-rpidistro to fix WiFi on Raspberry Pi 4.
This is a temporary solution while AGL master is still based on
Thud. As soon as AGL moves to newer release of the Yocto Project
and layer meta-raspberrypi this patch will become reduntant.

The change of boardflags3 value from "0x48200100" to "0x44200100"
has been done already in Raspbian (2019-06-20). For more details:
https://github.com/RPi-Distro/firmware-nonfree/issues/3

Bug-AGL: SPEC-2656

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>
Change-Id: Ifaafbae6ff59a43f95926e64f1ce35aefadd3cfa