meta-agl-profile-core: update curl and nghttp2
Backport newer curl and nghttp2 recipes, and bbappend curl to enable
nghttp2 support by default when building for target. In concert with
a change to mask out the curl bbappend in meta-aac, this should
provide the requirements for alexa-voiceagent-service and avoid
significant rebuilds when the agl-voiceagent-alexa feature is
enabled/disabled.

The curl recipe backport is from commit 0b7320b in upstream oe-core
master branch, the nghttp2 recipe backport is from commit 5a9b6d3 in
meta-networking in meta-openembedded master branch.

Bug-AGL: SPEC-2839

Change-Id: Id1ec872dfcb4ecc7a9ba68bdefb114ceeff066c3
Signed-off-by: Scott Murray <scott.murray@konsulko.com>