af-main: Sign widgets when agl-devel
This change introduce signature of widgets with
sample keys and certificates of app-framework-main
when the feature agl-devel is selected.

It also shows how to sign widgets in yocto build
environment, using WGTPKG_AUTOSIGN_X environment
variables.

v2: Use an own agl feature for the signing and not agl-devel
    This allows us to switch it on selectively.

Bug-AGL: SPEC-2840

Change-Id: I8f66c25a2ae62ad2365c7cdb995049c00b7c780a
Signed-off-by: José Bollo <jose.bollo@iot.bzh>