meta-agl-distro: BBMASK problematic upstream meta-security bbappend
Mask upstream meta-security bbappend that interferes with non-linux-yocto
kernels due to the bug of having two wildcards, which bitbake does not
support. The result is the bbappend applying to all linux-% recipes,
breaking e.g. linux-raspberry compilation.  Not having it is not an issue,
since there is also explicit enabling of required security features done
by meta-agl/meta-security.

Bug-AGL: SPEC-2932

Change-Id: I60579cf539c0b134dda06af75f98a7960330b147
Signed-off-by: Scott Murray <scott.murray@konsulko.com>