templates/feature/agl-compositor: Allow to test experimental agl-compositor
Facilitates testing the client shell (homescreen) with the newer
agl-compositor, but only enabled when DISTRO_FEATURES has
'agl-compositor'. Uses a WIP branch for it.

Regular clients, like hvac, mediaplayer, launcher and navigation-demo
also use the WIP branch. This modifies the conf/local file so it either
needs a clean build directory or it requires an override of the
conf/local when sourcing the aglsetup script.

Bug-AGL: SPEC-3123, SPEC-3124

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Suggested-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Change-Id: Ice650b5b61ebd1e1a35f045301c14e0bedda3a05