meta-agl-profile-core: tweak udisks2 configuration for AGL
Per discussion on Arch wiki at:

https://wiki.archlinux.org/index.php/Udisks#Mount_to_/media_(udisks2)

add a udev rules file to tweak the UDISKS_FILESYSTEM_SHARED environment
variable to configure udisks2 to use /media as used to be the default
with the older udisks. This restores behavior to what is expected by
the AGL configuration for lightmediascanner.

Bug-AGL: SPEC-2932

Change-Id: I66fda10f71dea889f3cd06341d9daa809bec2aa5
Signed-off-by: Scott Murray <scott.murray@konsulko.com>