meta-agl-profile-cluster-qt5: disable qtbase patches
The patches to qtbase for supporting the qtcompositor cluster demo
on dra7xx-evm do not apply against 5.13.2 and it is not clear if
they are still required, disable them for now until someone with
the hardware can investigate. Note that this change is required to
unblock building the regular cluster demo.

Bug-AGL: SPEC-2932

Change-Id: I47b7de17a0d2da5155b2c58f89bbe78ab516c4e1
Signed-off-by: Scott Murray <scott.murray@konsulko.com>