aglwgt.bbclass: remove pushd/popd usage
The recent addition of multiple widget types in aglwgt.bbclass added
usage of pushd and popd, which are bash specific. This is undesirable,
since it causes failures for anyone trying to build on default installs
of Debian or Ubuntu. Since the B and S variables in OE recipes are
known to be absolute paths by definition, the pushd/popd can simply
be replaced with "cd".

Bug-AGL: SPEC-2049

Change-Id: I7ec9c558545be9aedd0b7e40539127b7f8c665ed
Signed-off-by: Scott Murray <scott.murray@konsulko.com>