Remove network config fragment deployed by upstream
Upstream YP/zeus unconditionally deploys a network config fragment for
systemd-networkd.
This breaks our CI builds as it races with connman and netboot.
At least it does disturb the network enough for netboot to fail.

We have a nicer solution for that in-tree that we will upstream.
Until then, we remove the config file from our filesystems.

Bug-AGL: SPEC-3233
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Change-Id: I9e8343e77af378c058702b13fe000a344e6a1ef9