meta-agl-profile-graphical: add SMACK labelling for weston
The initial patch to allow disabling memfd usage in weston has
proven to be naive, as the v7 wayland seat resource changes in
Weston 8.0.0 are dependent on them.  To avoid needing to make more
invasive changes such as forcing the seat resource version back to
v6, drop the patch in favor of having Weston run under a
System::Weston SMACK label and adding the rules required to have it
work.

As well, use-XDG_RUNTIMESHARE_DIR.patch and the associated service
unit changes have been removed since they are not required now with
explicit labelling in place.

Bug-AGL: SPEC-3305, SPEC-3350

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I8aef287219a7f95992a82f4ec2ee8e1822ca4ce8