Use upstream meta-renesas from github and update meta-rcar-gen3-adas integration
Thanks to the work done by Renesas and members of our community, we can
now directly use meta-renesas from upstream github.
A shout-out to everyone involved (Kurokawa-san, Ronan, Stephane).

It turned-out it is simpler to create a symlink-based slim version
of meta-rcar-gen3-adas layer instead of BBMASK'ing what should
not be in there. See the comments in
meta-agl-bsp/meta-rcar-gen3-adas/conf/NOTES
as to why.

Remaining work is to include the packages outside of the layer.conf
in a follow-up.

Bug-AGL: SPEC-3380

Change-Id: I2be1b44049d74146546380e2eb866f0558a26822
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/24796
Reviewed-by: Stéphane Desneux <stephane.desneux@iot.bzh>